pr feedback: use tmp director instead of spec fixtures,  move capybara support out of rails_helper and into a support file,
